+++ python/lsst/ts/m2gui/model.py
@@ -1,6 +1,6 @@
 # This file is part of ts_m2gui.
 #
-# Developed for the LSST Data Management System.
+# Developed for the LSST Telescope and Site Systems.
 # This product includes software developed by the LSST Project
 # (https://www.lsst.org).
 # See the COPYRIGHT file at the top-level directory of this distribution
@@ -21,7 +21,18 @@
 
 __all__ = ["Model"]
 
-from lsst.ts.m2com import ActuatorDisplacementUnit
+import asyncio
+import time
+
+from lsst.ts import salobj
+from lsst.ts.m2com import (
+    ActuatorDisplacementUnit,
+    Controller,
+    MockServer,
+    get_config_dir,
+)
+from lsst.ts.tcpip import LOCAL_HOST
+from lsst.ts.utils import index_generator, make_done_future
 
 from . import (
     Config,
@@ -53,6 +64,9 @@ class Model(object):
         Telemetry port to connect. (the default is 50011)
     timeout_connection : `int` or `float`, optional
         Connection timeout in second. (the default is 10)
+    timeout_in_second : `float`, optional
+        Time limit for reading data from the TCP/IP interface (sec). (the
+        default is 0.05)
     is_simulation_mode: `bool`, optional
         True if running in simulation mode. (the default is False)
 
@@ -88,7 +102,18 @@ class Model(object):
         Telemetry port to connect.
     timeout_connection : `int` or `float`
         Connection timeout in second.
+    timeout_in_second : `float`
+        Time limit for reading data from the TCP/IP interface (sec).
+    controller : `lsst.ts.m2com.Controller`
+        Controller to do the TCP/IP communication with the servers of M2 cell.
+    stop_loop_timeout : `float`
+        Timeout of stoping loop in second.
+    run_loops : `bool`
+        The event and telemetry loops are running or not.
     """
+
+    # Maximum timeout to wait the telemetry in second
+    TELEMETRY_WAIT_TIMEOUT = 900
 
     def __init__(
         self,
@@ -97,6 +122,7 @@ class Model(object):
         port_command=50010,
         port_telemetry=50011,
         timeout_connection=10,
+        timeout_in_second=0.05,
         is_simulation_mode=False,
     ):
 
@@ -123,7 +149,31 @@ class Model(object):
         self.port_telemetry = port_telemetry
         self.timeout_connection = timeout_connection
 
+        self.timeout_in_second = timeout_in_second
+        self.controller = Controller(
+            log=self.log, timeout_in_second=self.timeout_in_second
+        )
+
         self._is_simulation_mode = is_simulation_mode
+
+        # Mock server that is only needed in the simulation mode
+        self._mock_server = None
+
+        self.stop_loop_timeout = 5.0
+
+        # Sequence generator
+        self._sequence_generator = index_generator()
+
+        self.run_loops = False
+
+        # Task of the telemetry loop from component (asyncio.Future)
+        self._task_telemetry_loop = make_done_future()
+
+        # Task of the event loop from component (asyncio.Future)
+        self._task_event_loop = make_done_future()
+
+        # Task to monitor the connection status actively (asyncio.Future)
+        self._task_connection_monitor_loop = make_done_future()
 
     def _set_system_status(self):
         """Set the default system status.
@@ -530,17 +580,22 @@ class Model(object):
         ------
         `RuntimeError`
             There is the connection with the M2 controller already.
+        `ValueError`
+            When the command port equals to the telemetry port.
         """
 
         if self.system_status["isCrioConnected"]:
             raise RuntimeError("Please disconnect from the M2 controller first.")
+
+        if port_command == port_telemetry:
+            raise ValueError("Command port should be different from telemetry port.")
 
         self.host = host
         self.port_command = port_command
         self.port_telemetry = port_telemetry
         self.timeout_connection = timeout_connection
 
-    def connect(self):
+    async def connect(self):
         """Connect to the M2 controller.
 
         Raises
@@ -552,18 +607,285 @@ class Model(object):
         if self.system_status["isCrioConnected"]:
             raise RuntimeError("Please disconnect from the M2 controller first.")
 
-        self.log.info("Connect to the M2 controller.")
+        self.log.info("Connecting to the M2 controller...")
 
-    def disconnect(self):
-        """Disconnect from the M2 controller.
+        # Run the mock server in the simulation mode
+        if self._is_simulation_mode:
+            await self._run_mock_server()
+
+        # Run the event and telemetry loops
+        if self.run_loops is False:
+
+            self.run_loops = True
+
+            self.log.debug(
+                "Starting event, telemetry and connection monitor loop tasks."
+            )
+
+            self._task_event_loop = asyncio.create_task(self._event_loop())
+            self._task_telemetry_loop = asyncio.create_task(self._telemetry_loop())
+            self._task_connection_monitor_loop = asyncio.create_task(
+                self._connection_monitor_loop()
+            )
+
+        await self._connect_server()
+
+        self.update_system_status(
+            "isCrioConnected", self.controller.are_clients_connected()
+        )
+
+    async def _run_mock_server(self):
+        """Run the mock server to support the simulation mode."""
+
+        # Close the running mock server if any
+        if self._mock_server is not None:
+            await self._mock_server.close()
+            self._mock_server = None
+
+        # Run a new mock server
+        self._mock_server = MockServer(
+            LOCAL_HOST,
+            port_command=0,
+            port_telemetry=0,
+            log=self.log,
+            is_csc=False,
+        )
+        self._mock_server.model.configure(get_config_dir(), "harrisLUT")
+        await self._mock_server.start()
+
+    async def _event_loop(self):
+        """Update and output event information from component."""
+
+        self.log.debug("Begin to run the event loop from component.")
+
+        while self.run_loops:
+
+            try:
+                message = (
+                    self.controller.queue_event.get_nowait()
+                    if not self.controller.queue_event.empty()
+                    else await self.controller.queue_event.get()
+                )
+
+            # When there is no instance in the self.controller.queue_event, the
+            # get() will be used to wait for the new coming event. It is a
+            # blocking call and will get the asyncio.CancelledError if we
+            # cancel the self._task_event_loop.
+            except asyncio.CancelledError:
+                message = ""
+
+            self._process_events(message)
+
+            # Transition to the Diagnostic state if the controller is in Fault
+            if (self.local_mode == LocalMode.Enable) and (
+                self.controller.controller_state == salobj.State.FAULT
+            ):
+                # TODO: Implement this in DM-36015
+                pass
+
+            else:
+                await asyncio.sleep(self.timeout_in_second)
+
+        self.log.debug("Component's event loop exited.")
+
+    def _process_events(self, message):
+        """Process the events from the M2 controller.
+
+        Parameters
+        ----------
+        message : `str`
+            Message from the M2 controller.
+        """
+        # TODO: Implement this in DM-36015
+        pass
+
+    async def _telemetry_loop(self):
+        """Update and output telemetry information from component."""
+
+        self.log.debug("Starting telemetry loop from component.")
+
+        time_wait_telemetry = 0.0
+        is_telemetry_timed_out = False
+
+        while self.run_loops:
+
+            if self.controller.are_clients_connected():
+
+                if (
+                    time_wait_telemetry >= self.TELEMETRY_WAIT_TIMEOUT
+                    and not is_telemetry_timed_out
+                ):
+
+                    self.log.warning(
+                        (
+                            "No telemetry update for more than "
+                            f"{self.TELEMETRY_WAIT_TIMEOUT} seconds. Can you "
+                            f"ping the controller at {self.host} and initiate "
+                            "connection to one of it's ports (at "
+                            f"{self.port_command} and {self.port_telemetry})?"
+                        )
+                    )
+                    time_wait_telemetry = 0.0
+                    is_telemetry_timed_out = True
+
+                # If there is no telemetry, sleep for some time
+                if self.controller.client_telemetry.queue.empty():
+                    await asyncio.sleep(self.timeout_in_second)
+                    time_wait_telemetry += self.timeout_in_second
+                    continue
+
+                # There is the telemetry to publish
+                time_wait_telemetry = 0.0
+                if is_telemetry_timed_out:
+                    self.log.info("Telemetry is up and running after failure.")
+
+                is_telemetry_timed_out = False
+
+                message = self.controller.client_telemetry.queue.get_nowait()
+
+                # Send the signals of telemetry
+                self._process_telemetry(message)
+
+            else:
+                self.log.debug(
+                    f"Clients not connected. Waiting {self.timeout_in_second}s..."
+                )
+                await asyncio.sleep(self.timeout_in_second)
+
+        self.log.debug("The component's telemetry loop exited/ended.")
+
+    def _process_telemetry(self, message):
+        """Process the telemetry from the M2 controller.
+
+        Parameters
+        ----------
+        message : `str`
+            Message from the M2 controller.
+        """
+        # TODO: Implement this in DM-36015
+        pass
+
+    async def _connection_monitor_loop(self, period=1):
+        """Actively monitor the connection status from component. Disconnect
+        the system if the connection is lost by itself.
+
+        Parameters
+        ----------
+        period : `float` or `int`, optional
+            Period to check the connection status in second. (the default is 1)
+        """
+
+        self.log.debug("Begin to run the connection monitor loop from component.")
+
+        were_clients_connected = False
+        while self.run_loops:
+
+            if self.controller.are_clients_connected():
+                were_clients_connected = True
+            else:
+                if were_clients_connected:
+                    were_clients_connected = False
+                    self.log.info(
+                        "Lost the TCP/IP connection in the active monitoring loop."
+                    )
+
+                    await self.controller.close()
+                    self.update_system_status(
+                        "isCrioConnected", self.controller.are_clients_connected()
+                    )
+
+            await asyncio.sleep(period)
+
+        self.log.debug("Connection monitor loop from component closed.")
+
+    async def _connect_server(self):
+        """Connect the TCP/IP server.
 
         Raises
         ------
-        `RuntimeError`
-            There is no connection with the M2 controller yet.
+        RuntimeError
+            If timeout in connection.
         """
 
-        if not self.system_status["isCrioConnected"]:
-            raise RuntimeError("There is no connection with the M2 controller.")
+        port_command = self.port_command
+        port_telemetry = self.port_telemetry
 
-        self.log.info("Disconnect from the M2 controller.")
+        # In the simulation mode, the ports of mock server are randomly
+        # assigned by the operation system.
+        if self._is_simulation_mode:
+            port_command = self._mock_server.server_command.port
+            port_telemetry = self._mock_server.server_telemetry.port
+
+        self.log.debug(f"Host in connection request: {self.host}")
+        self.log.debug(f"Command port in connection request: {port_command}")
+        self.log.debug(f"Telemetry port in connection request: {port_telemetry}")
+
+        self.controller.start(
+            self.host,
+            port_command,
+            port_telemetry,
+            sequence_generator=self._sequence_generator,
+            timeout=self.timeout_connection,
+        )
+
+        time_start = time.monotonic()
+        connection_pooling_time = 0.1
+        while not self.controller.are_clients_connected() and (
+            (time.monotonic() - time_start) < self.timeout_connection
+        ):
+            await asyncio.sleep(connection_pooling_time)
+
+        if not self.controller.are_clients_connected():
+            raise RuntimeError(
+                "Timeout in connection - Conection timeouted, connection "
+                f"failed or cannot connect. Host: {self.host}, ports: "
+                f"{port_command} and {port_telemetry}."
+            )
+
+    async def disconnect(self):
+        """Disconnect from the M2 controller."""
+
+        if self.controller.are_clients_connected():
+            self.log.info("Disconnecting from the M2 controller...")
+        else:
+            self.log.info("No connection with the M2 controller.")
+
+        await self.close_tasks()
+        self.update_system_status(
+            "isCrioConnected", self.controller.are_clients_connected()
+        )
+
+    async def close_tasks(self):
+        """Close the asynchronous tasks."""
+
+        await self.controller.close()
+
+        if self._mock_server is not None:
+            # Wait some time to let the mock server notices the controller has
+            # closed the connection.
+            await asyncio.sleep(10)
+
+            await self._mock_server.close()
+            self._mock_server = None
+
+        try:
+            await self._stop_loops()
+        except Exception:
+            self.log.exception("Exception while stopping the loops. Ignoring...")
+
+    async def _stop_loops(self):
+        """Stop the loops."""
+
+        self.run_loops = False
+
+        for task in (
+            self._task_telemetry_loop,
+            self._task_event_loop,
+            self._task_connection_monitor_loop,
+        ):
+            try:
+                await asyncio.wait_for(task, timeout=self.stop_loop_timeout)
+            except asyncio.TimeoutError:
+                self.log.debug("Timed out waiting for the loop to finish. Canceling.")
+
+                task.cancel()